+++ hack/helm/release/main.go
@@ -1,7 +1,7 @@
-/*
-Copyright © 2022 NAME HERE <EMAIL ADDRESS>
+// Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+// or more contributor license agreements. Licensed under the Elastic License 2.0;
+// you may not use this file except in compliance with the Elastic License 2.0.
 
-*/
 package main
 
 import "github.com/elastic/cloud-on-k8s/hack/helm/release/cmd"
