+++ hack/helm/release/readme.md
@@ -1,14 +1,14 @@
 # TODO
 
 - [x] Handle updating index
-- [ ] Add dry run?
-- [ ] Excludes
-- [ ] Release single chart?
+- [x] Add dry run?
+- [x] Excludes
+- [x] Release single chart?
 - [ ] New logger with info/debug capabilities?
-- [ ] Tests
-- [ ] Manual Testing
-- [ ] Better organization, and naming of funcs
-- [ ] Remove writing to fake gcs bucket
+- [x] Tests
+- [x] Manual Testing
+- [x] Better organization, and naming of funcs
+- [x] Remove writing to fake gcs bucket
 - [ ] Validation of chart repo url + bucket combinations
 - [ ] Documentation
 - [ ] Personal review
